out/production/professor/fastapi.sample/requirements/base.txt
alembic
bcrypt
fastapi
httpx
psycopg2-binary
pydantic[email]
redis
requests
SQLALchemy
sqlalchemy_continuum
tenacity
python-keycloak
PyJWT